Separated parsers for individual files
