+++ simphony_metaedit/parsers/exceptions.py
@@ -1,0 +1,3 @@
+class ParsingError(Exception):
+    """Raised if an anomalous conditions is found during parsing or linkage"""
+    pass
